ikev2 ipseckey: split out DNS query code into a separate file
When compiled with DNSSEC support, <ldns/ldns.h> indirectly pulls in
OpenSSL headers where symbols conflicting with the ones defined in NSS
headers, such as KU_* macros.  This moves out the low-level DNS lookup
code using LDNS to a dedicated file, so there will be no interference.
