+++ programs/pluto/Makefile
@@ -210,7 +210,7 @@ OBJS += ikev1_retry.o
 OBJS += ikev2_retry.o
 OBJS += ipsec_doi.o
 ifeq ($(USE_DNSSEC),true)
-OBJS += ikev2_ipseckey.o
+OBJS += ikev2_ipseckey.o ikev2_ipseckey_dnsr.o
 endif
 ifeq ($(USE_IKEv1),true)
 OBJS += ikev1.o ikev1_aggr.o ikev1_main.o ikev1_quick.o ikev1_dpd.o ikev1_spdb_struct.o ikev1_msgid.o
