+++ programs/pluto/ikev2_ipseckey.h
@@ -8,8 +8,5 @@
 
 #define IS_LIBUNBOUND LSW_LIBUNBOUND_ENABLED
 
-struct p_dns_req;
-
 extern stf_status idi_ipseckey_fetch(struct ike_sa *ike);
-extern void free_ipseckey_dns(struct p_dns_req *pp);
 extern stf_status idr_ipseckey_fetch(struct ike_sa *ike);
